seperated individuals from system config
